Add core module to the project
there is the need for a new module containing all game core concepts,
for example Tower package or Card Package does not rely on server or
client and on the other hand both server and client need to use
these packages and some how these package is a common dependency of
both of this modules.
So this module will contain all those packages and their basic implementation
that does not rely on the Server or the Client, and then with maven we make a jar file
of the core module and then add it as a dependency to both Server and Client module.