Refactor abstract tower class to just shoot target
the shoot method before had a shoot method that in it, a timer for
one seconds had to gone by and then it will reduce the health of the
attack target, but the shoot target should not be responsible for
doing the counter, the client code must count the timer and then
call this method. so refactored the method to just reduce the target
health when called by the damage count.