Fix no constructor problem for Card class
there was the problem that Card base class didn't have a constructor
and so didn't force the Troop, Buildings, Spell, class to have a
constructor and specify the id, cost, owner, and position of the
Card.
So refactored it to have a constructor for setting those values and then
forcing children to specify those values.