Refactor Troop class to have target and self type
there is the need for every troop to know it's own type.for example
a barberian is a ground type of troop and baby dragon dragon is a
air type of troop.
Some troops can't attack air targets and have to just go pass that
target.
so the ability for checking the incoming troop target has been added.