Implemented singleton pattern for Connector class
the connector class needs to be accessible from all places of the code.
So it must implement the singleton pattern, but the problem is that it's controller
have 3 arguments, so had to come up with another function called connect,
which will be called at project bootstrap, and then in all other parts of the
project if we call it we will get Excepiton, for getting the instance
of the connection client must use the getInstance() method.