Refactor Authenication Response class to store User
the authenication response commands should store a user inside of it
to be able to send it through the socket and it's Input & Output streams
to the client, so the client can store it, and use it as it wishes.
chagned the classes to have two classes, one for when their response
is sucessful, and one for when the response is not sucessful.
when response is sucessful it will store a user.
if it isn't sucessful it will have a null pointer in it's user field.