Add UserData to global classes.
the global classes will contain data related to all controllers and
pages. in order to be able to store the data the UserData class
have a public static feild called User which will be set to User
recived from the login or register response commands.
it will contain all the data related to the user.