Refactor Card class to implement KnowsItsExactType
there is the need for every card to know it's exact type.
so it is better to have a mechanism in the Card class to handle this
checks.
This mechanism is implemented using KnowsItsExactTypeTrait which
extends three other traits which are : KnowsItsExactTroopTypeTrait,
KnowsItsExactBuildingTypeTrait, KnowsItsExactSpelltypeTrait.
which all of them have default methods that return a boolean
indicating if they are a building, troop or a spell.