Fix bug in isDead() function of attackables
there was a bug in isDead() function of attackables that the
function assumed that a card will reach hp zero, but there
was the possibility that a card is attacked and it health get's
lower than zero and becomes negative in that single attack.
so it must declare the target as killed.